trainer: Move reward fn logic into discrim_net
